refactor: factorize description DOM insertion logic in updateDescriptionElement
